release: ship v0.9.9 vulnerability repair flow

+++ crates/synaptic-api/src/repair.rs
@@ -52,7 +52,7 @@ pub struct GateResult {
     pub gate: String,
     pub outcome: GateOutcome,
     pub detail: String,
-    pub duration_ms: u128,
+    pub duration_ms: u64,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
